Move anthem.cpp to new folder and adjust imports
