+++ engine/src/command_handlers/project_command_handler.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "messages_generated.h"
-#include "../anthem.h"
+#include "anthem.h"
 #include "../plugin_window.h"
 
 std::optional<flatbuffers::Offset<Response>> handleProjectCommand(const Request* request, flatbuffers::FlatBufferBuilder& builder, Anthem* anthem);
