+++ engine/src/modules/core/anthem.cpp
